Uploaded v1 of the Main java file
This program can show the user the appointment codes for the current day, and also allows an appointment to be "booked" by clicking a button.
This is very basic, but shows the customer what is possible. Further iterations would keep a log of appointments that have been booked and offer the user the nearest non booked appointment, give the user an option to decline an appointment, and eventually allow the user to book a time slot directly. However this is a good start to show Beatrice.